refactor: Adapterを別ファイルに切り出した
